+++ lib/analytics/consumers/consumer.php
@@ -3,6 +3,7 @@ abstract class Analytics_Consumer {
 
   protected $options;
   protected $secret;
+  protected $type = "Consumer";
 
   /**
    * Store our secret and options as part of this consumer
@@ -40,6 +41,35 @@ abstract class Analytics_Consumer {
    */
   protected function debug() {
     return isset($this->options["debug"]) ? $this->options["debug"] : false;
+  }
+
+  /**
+   * Check whether we should connect to the API using SSL. This is enabled by
+   * default with connections which make batching requests. For connections
+   * which can save on round-trip times, we disable it.
+   * @return boolean
+   */
+  protected function ssl() {
+    return isset($this->options["ssl"]) ? $this->options["ssl"] : false;
+  }
+
+
+  /**
+   * On an error, try and call the error handler, if debugging output to
+   * error_log as well.
+   * @param  string $code
+   * @param  string $msg
+   */
+  protected function handleError($code, $msg) {
+
+    if (isset($this->options['error_handler'])) {
+      $handler = $this->options['error_handler'];
+      $handler($code, $msg);
+    }
+
+    if ($this->debug()) {
+      error_log("[Analytics][" . $this->type . "] " . $msg);
+    }
   }
 }
 ?>